feat(storage): Firebase JS SDK v9 modular API (#6958)
