+++ tests/app.js
@@ -38,6 +38,7 @@ import * as perfModular from '@react-native-firebase/perf';
 import '@react-native-firebase/remote-config';
 import * as remoteConfigModular from '@react-native-firebase/remote-config';
 import '@react-native-firebase/storage';
+import * as storageModular from '@react-native-firebase/storage';
 import jet from 'jet/platform/react-native';
 import React from 'react';
 import { AppRegistry, Button, NativeModules, Text, View } from 'react-native';
@@ -54,6 +55,7 @@ jet.exposeContextProperty('remoteConfigModular', remoteConfigModular);
 jet.exposeContextProperty('perfModular', perfModular);
 jet.exposeContextProperty('appCheckModular', appCheckModular);
 jet.exposeContextProperty('messagingModular', messagingModular);
+jet.exposeContextProperty('storageModular', storageModular);
 
 firebase.database().useEmulator('localhost', 9000);
 firebase.auth().useEmulator('http://localhost:9099');
